modproxy: rename s3proxy.Cacher to modproxy.S3Cacher
